默认的

+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/SysApiRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface SysApiRepository extends JpaRepository<SysApi, Integer>, JpaSpecificationExecutor {
+public interface SysApiRepository extends JpaRepository<SysApi, Long>, JpaSpecificationExecutor {
 
 }
