+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/SysBlacklistRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface SysBlacklistRepository extends JpaRepository<SysBlacklist, Integer>, JpaSpecificationExecutor {
+public interface SysBlacklistRepository extends JpaRepository<SysBlacklist, Long>, JpaSpecificationExecutor {
 
 }
