+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/SysMenuRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface SysMenuRepository extends JpaRepository<SysMenu, Integer>, JpaSpecificationExecutor {
+public interface SysMenuRepository extends JpaRepository<SysMenu, Long>, JpaSpecificationExecutor {
 
 }
