+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/WccAttachmentRepository.java
@@ -4,13 +4,11 @@ import hqsc.ray.wcc2.entity.WccAttachment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.util.List;
-
 /**
  * 描述：
  *
  * @author Administrator
  */
-public interface WccAttachmentRepository extends JpaRepository<WccAttachment, Integer>, JpaSpecificationExecutor {
+public interface WccAttachmentRepository extends JpaRepository<WccAttachment, Long>, JpaSpecificationExecutor {
 
 }
