+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/WccCircleInfoRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface WccCircleInfoRepository extends JpaRepository<WccCircleInfo, Integer>, JpaSpecificationExecutor {
+public interface WccCircleInfoRepository extends JpaRepository<WccCircleInfo, Long>, JpaSpecificationExecutor {
 
 }
