+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/WccMoneyReceiptsRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface WccMoneyReceiptsRepository extends JpaRepository<WccMoneyReceipts, Integer>, JpaSpecificationExecutor {
+public interface WccMoneyReceiptsRepository extends JpaRepository<WccMoneyReceipts, Long>, JpaSpecificationExecutor {
 
 }
