+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/WccOpenVipPriceRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface WccOpenVipPriceRepository extends JpaRepository<WccOpenVipPrice, Integer>, JpaSpecificationExecutor {
+public interface WccOpenVipPriceRepository extends JpaRepository<WccOpenVipPrice, Long>, JpaSpecificationExecutor {
 
 }
