+++ ray-platform/ray-wcc2/src/main/java/hqsc/ray/wcc2/repository/WccUserCircleRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
  *
  * @author Administrator
  */
-public interface WccUserCircleRepository extends JpaRepository<WccUserCircle, Integer>, JpaSpecificationExecutor {
+public interface WccUserCircleRepository extends JpaRepository<WccUserCircle, Long>, JpaSpecificationExecutor {
 
 }
